web_interface/backend/requirements.txt
# Qubic Backend Dependencies
flask~=3.0
flask-cors~=4.0
gunicorn~=21.0
python-dotenv~=1.0
trimesh~=4.0
numpy~=1.24
networkx~=3.0
flask-limiter~=3.5
PyJWT[crypto]>=2.8.0
